requirements.txt
requests==2.28.1
Pillow==9.4.0
matplotlib==3.7.1
pandas==1.5.3
numpy==1.24.2